src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/content-pseudo-element-not-last-expected.txt

CONSOLE ERROR: line 9: Uncaught TypeError: host.createShadowRoot is not a function
Accept simple selectors after ::content For Polymer 0.5 compat

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE

This text should be green
